requirements.txt
# Basic packages
numpy==1.26.4
pandas==2.2.2
tabulate==0.9.0
matplotlib==3.8.4
seaborn==0.13.2
Pillow>=10.0.0

# Core DL framework
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
torchmetrics[image]>=1.3.0

# Detection / models
ultralytics==8.3.0
opencv-python>=4.8.0
transformers>=4.38,<5
timm>=0.9.12
